+++ packages/amplify_flutter/lib/categories/amplify_storage_category.dart
@@ -39,10 +39,13 @@ class StorageCategory {
   }
 
   Future<UploadFileResult> uploadFile(
-      {required File local, required String key, UploadFileOptions? options}) {
+      {required File local,
+      required String key,
+      void Function(TransferProgress)? onProgress,
+      UploadFileOptions? options}) {
     final UploadFileRequest request =
         UploadFileRequest(local: local, key: key, options: options);
-    return plugins[0].uploadFile(request: request);
+    return plugins[0].uploadFile(request: request, onProgress: onProgress);
   }
 
   Future<GetUrlResult> getUrl({required String key, GetUrlOptions? options}) {
@@ -63,9 +66,10 @@ class StorageCategory {
   Future<DownloadFileResult> downloadFile(
       {required String key,
       required File local,
+      void Function(TransferProgress)? onProgress,
       DownloadFileOptions? options}) {
     final DownloadFileRequest request =
         DownloadFileRequest(key: key, local: local, options: options);
-    return plugins[0].downloadFile(request: request);
+    return plugins[0].downloadFile(request: request, onProgress: onProgress);
   }
 }