+++ packages/amplify_storage_s3/example/integration_test/main_test.dart
@@ -76,8 +76,23 @@ void main() async {
       metadata['desc'] = 'A test file';
       S3UploadFileOptions options = S3UploadFileOptions(
           accessLevel: StorageAccessLevel.guest, metadata: metadata);
+
+      int fileLength = file.lengthSync();
+      int lastProgressBytes = 0;
+      int onProgressCalls = 0;
       final result = await Amplify.Storage.uploadFile(
-          key: key, local: file, options: options);
+          key: key,
+          local: file,
+          options: options,
+          onProgress: (progress) {
+            expect(
+                progress.currentBytes, greaterThanOrEqualTo(lastProgressBytes));
+            lastProgressBytes = progress.currentBytes;
+
+            expect(progress.totalBytes, fileLength);
+            onProgressCalls++;
+          });
+      expect(onProgressCalls, greaterThan(0));
       lastUploadedKey = result.key;
       expect(lastUploadedKey, key);