+++ packages/amplify_storage_s3/ios/Classes/SwiftAmplifyStorageS3Plugin.swift
@@ -21,14 +21,19 @@ import AWSMobileClient
 import amplify_core
 
 public class SwiftAmplifyStorageS3Plugin: NSObject, FlutterPlugin {
+    
+    private static var transferProgressStreamHandler : TransferProgressStreamHandler = TransferProgressStreamHandler()
+    
     public static func register(with registrar: FlutterPluginRegistrar) {
         let channel = FlutterMethodChannel(name: "com.amazonaws.amplify/storage_s3", binaryMessenger: registrar.messenger())
         let instance = SwiftAmplifyStorageS3Plugin()
         registrar.addMethodCallDelegate(instance, channel: channel)
+        
+        FlutterEventChannel(name: "com.amazonaws.amplify/storage_transfer_progress_events", binaryMessenger: registrar.messenger()).setStreamHandler(transferProgressStreamHandler)
     }
 
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
-    print("In handle for method \(call.method)")
+    let result = AtomicResult(result, call.method)
 
     if(call.method == "addPlugin"){
         do {
@@ -73,7 +78,7 @@ public class SwiftAmplifyStorageS3Plugin: NSObject, FlutterPlugin {
     let arguments = call.arguments as! Dictionary<String,AnyObject>
     switch call.method {
         case "uploadFile":
-            AmplifyStorageOperations.uploadFile(flutterResult: result, request: arguments)
+            AmplifyStorageOperations.uploadFile(flutterResult: result, request: arguments, transferProgressStreamHandler: SwiftAmplifyStorageS3Plugin.transferProgressStreamHandler)
         case "getUrl":
             AmplifyStorageOperations.getURL(flutterResult: result, request: arguments)
         case "remove":
@@ -81,7 +86,7 @@ public class SwiftAmplifyStorageS3Plugin: NSObject, FlutterPlugin {
         case "list":
             AmplifyStorageOperations.list(flutterResult: result, request: arguments)
         case "downloadFile":
-            AmplifyStorageOperations.downloadFile(flutterResult: result, request: arguments)
+            AmplifyStorageOperations.downloadFile(flutterResult: result, request: arguments, transferProgressStreamHandler: SwiftAmplifyStorageS3Plugin.transferProgressStreamHandler)
         default:
             result(FlutterMethodNotImplemented)
     }