+++ packages/amplify_auth_cognito/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.2.7 (2021-11-08)
+
+### Chores
+
+- chore: Bump Amplify iOS to 1.15.5
+
+## 0.2.6 (2021-10-25)
+
+## 0.2.5 (2021-10-14)
+
+### Fixes
+- fix(auth): Add global sign out
+- fix(auth): Support `preferPrivateSession` flag
+
 ## 0.2.4 (2021-09-10)
 
 ### Fixes